refactor: 카드 추가 화면을 카드 입력 화면으로 이름 변경
