+++ app/src/androidTest/java/woowacourse/payments/ui/features/cartinput/components/CardPasswordFieldTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.features.addcard.components
+package woowacourse.payments.ui.features.cartinput.components
 
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
