+++ app/src/androidTest/java/woowacourse/payments/ui/features/cartinput/components/bottomsheet/BottomSheetCardCompanyItemTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.features.addcard.components.bottomsheet
+package woowacourse.payments.ui.features.cartinput.components.bottomsheet
 
 import androidx.activity.ComponentActivity
 import androidx.compose.ui.test.assertIsDisplayed
