+++ app/src/main/java/woowacourse/payments/AddcardActivity.kt
@@ -8,7 +8,7 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.result.ActivityResult
 import woowacourse.payments.domain.card.PaymentCard
-import woowacourse.payments.ui.features.addcard.AddCardScreen
+import woowacourse.payments.ui.features.cartinput.CardInputScreen
 import woowacourse.payments.ui.mapper.CardMapper.toUiModel
 import woowacourse.payments.ui.model.PaymentCardUiModel
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
@@ -20,7 +20,7 @@ class AddcardActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             AndroidpaymentsTheme {
-                AddCardScreen(
+                CardInputScreen(
                     onNavigateBack = { finish() },
                     onNavigateSave = { card: PaymentCard ->
                         val intent =
