+++ app/src/main/java/woowacourse/payments/ui/features/cartinput/components/CardPasswordField.kt
@@ -1,7 +1,5 @@
-package woowacourse.payments.ui.features.addcard.components
+package woowacourse.payments.ui.features.cartinput.components
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
@@ -15,67 +13,58 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
 import woowacourse.payments.R
-import woowacourse.payments.domain.card.values.CardNumber.Companion.MAX_LENGTH_CARD_NUMBER
+import woowacourse.payments.domain.card.values.Password.Companion.MAX_LENGTH_PASSWORD
 import woowacourse.payments.ui.components.AppTextField
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
-import woowacourse.payments.ui.util.SeparatorVisualTransformation
-
-const val CARD_NUMBER_CHUNK_SIZE = 4
-const val CARD_NUMBER_SEPARATOR = " - "
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun CardNumberField(
+fun CardPasswordField(
     modifier: Modifier = Modifier,
     value: String,
     onValueChange: (String) -> Unit,
-    supportingTextHeight: Dp = 20.dp,
 ) {
     var isFocused by remember { mutableStateOf(false) }
-    val isIncomplete = value.isNotEmpty() && value.length < MAX_LENGTH_CARD_NUMBER
+    val isIncomplete = value.isNotEmpty() && value.length < MAX_LENGTH_PASSWORD
     val showError = isIncomplete && !isFocused
 
-    val visualTransformation = remember { SeparatorVisualTransformation(CARD_NUMBER_CHUNK_SIZE, CARD_NUMBER_SEPARATOR) }
-
+    val visualTransformation = remember { PasswordVisualTransformation() }
     AppTextField(
         value = value,
         onValueChange = { newValue ->
-            val filteredValue = newValue.filter { it in '0'..'9' }.take(MAX_LENGTH_CARD_NUMBER)
+            val filteredValue = newValue.filter { it in '0'..'9' }.take(MAX_LENGTH_PASSWORD)
             onValueChange(filteredValue)
         },
         modifier =
             modifier.onFocusChanged { focusState ->
                 isFocused = focusState.isFocused
             },
-        labelText = stringResource(R.string.add_card_number_field_title),
-        placeholderText = stringResource(R.string.add_card_number_field_hint),
+        labelText = stringResource(R.string.add_card_password_field_title),
+        placeholderText = stringResource(R.string.add_card_password_field_hint),
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.NumberPassword),
         isError = showError,
         supportingText = {
-            Box(modifier = Modifier.height(supportingTextHeight)) {
-                if (showError) {
-                    Text(
-                        modifier = Modifier,
-                        text = stringResource(id = R.string.add_card_number_incomplete_error_message),
-                        color = MaterialTheme.colorScheme.error,
-                    )
-                }
+            if (showError) {
+                Text(
+                    modifier = Modifier,
+                    text = stringResource(id = R.string.add_card_password_incomplete_error_message),
+                    color = MaterialTheme.colorScheme.error,
+                )
             }
         },
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
         visualTransformation = visualTransformation,
     )
 }
 
 @Preview(showBackground = true)
 @Composable
-fun CardNumberFieldPreview() {
+fun CardPasswordFieldPreview() {
     var text by remember { mutableStateOf("") }
     AndroidpaymentsTheme(dynamicColor = false) {
-        CardNumberField(
+        CardPasswordField(
             value = text,
             onValueChange = { text = it },
         )
