+++ app/src/main/java/woowacourse/payments/ui/mapper/CardMapper.kt
@@ -9,12 +9,12 @@ import woowacourse.payments.domain.card.values.ExpireDate.Companion.create
 import woowacourse.payments.domain.card.values.OwnerName
 import woowacourse.payments.domain.card.values.Password
 import woowacourse.payments.ui.components.toMaskedString
-import woowacourse.payments.ui.features.addcard.CardUiState
-import woowacourse.payments.ui.features.addcard.ExpireDateUiState
-import woowacourse.payments.ui.features.addcard.components.CARD_NUMBER_CHUNK_SIZE
-import woowacourse.payments.ui.features.addcard.components.CARD_NUMBER_SEPARATOR
-import woowacourse.payments.ui.features.addcard.components.EXPIRE_DATE_CHUNK_SIZE
-import woowacourse.payments.ui.features.addcard.components.EXPIRE_DATE_SEPARATOR
+import woowacourse.payments.ui.features.cartinput.CardUiState
+import woowacourse.payments.ui.features.cartinput.ExpireDateUiState
+import woowacourse.payments.ui.features.cartinput.components.CARD_NUMBER_CHUNK_SIZE
+import woowacourse.payments.ui.features.cartinput.components.CARD_NUMBER_SEPARATOR
+import woowacourse.payments.ui.features.cartinput.components.EXPIRE_DATE_CHUNK_SIZE
+import woowacourse.payments.ui.features.cartinput.components.EXPIRE_DATE_SEPARATOR
 import woowacourse.payments.ui.model.CardCompanyUiModel
 import woowacourse.payments.ui.model.ExpireDateStatus
 import woowacourse.payments.ui.model.ExpireDateStatus.Invalid.ExpireDateInvalidReason
